backendd/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Werkzeug
#
scikit-learn
 numpy 
 pandas 
 pickle-mixin 
 twilio 
 google-generativeai